Move logging show utils to separate file
